backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
supabase==2.3.4
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
pytesseract==0.3.10
Pillow==10.1.0
joblib==1.3.2
regex==2023.10.3
statsmodels==0.14.1
scipy==1.11.4
opencv-python==4.8.1.78